+++ internal/fleet/integration_policy/delete.go
@@ -1,0 +1,29 @@
+package integration_policy
+
+import (
+	"context"
+
+	"github.com/elastic/terraform-provider-elasticstack/internal/clients/fleet"
+	"github.com/hashicorp/terraform-plugin-framework/resource"
+)
+
+func (r *integrationPolicyResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
+	var stateModel integrationPolicyModel
+
+	diags := req.State.Get(ctx, &stateModel)
+	resp.Diagnostics.Append(diags...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	client, err := r.client.GetFleetClient()
+	if err != nil {
+		resp.Diagnostics.AddError(err.Error(), "")
+		return
+	}
+
+	policyID := stateModel.PolicyID.ValueString()
+	force := stateModel.Force.ValueBool()
+	diags = fleet.DeletePackagePolicy(ctx, client, policyID, force)
+	resp.Diagnostics.Append(diags...)
+}