+++ provider/provider.go
@@ -10,7 +10,6 @@ import (
 	"github.com/elastic/terraform-provider-elasticstack/internal/elasticsearch/security"
 	"github.com/elastic/terraform-provider-elasticstack/internal/elasticsearch/transform"
 	"github.com/elastic/terraform-provider-elasticstack/internal/elasticsearch/watcher"
-	"github.com/elastic/terraform-provider-elasticstack/internal/fleet"
 	"github.com/elastic/terraform-provider-elasticstack/internal/kibana"
 	providerSchema "github.com/elastic/terraform-provider-elasticstack/internal/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
@@ -81,8 +80,6 @@ func New(version string) *schema.Provider {
 
 			"elasticstack_kibana_action_connector": kibana.DataSourceConnector(),
 			"elasticstack_kibana_security_role":    kibana.DataSourceRole(),
-
-			"elasticstack_fleet_integration": fleet.DataSourceIntegration(),
 		},
 		ResourcesMap: map[string]*schema.Resource{
 			"elasticstack_elasticsearch_cluster_settings":      cluster.ResourceSettings(),
@@ -109,12 +106,6 @@ func New(version string) *schema.Provider {
 			"elasticstack_kibana_action_connector": kibana.ResourceActionConnector(),
 			"elasticstack_kibana_security_role":    kibana.ResourceRole(),
 			"elasticstack_kibana_slo":              kibana.ResourceSlo(),
-
-			"elasticstack_fleet_agent_policy":       fleet.ResourceAgentPolicy(),
-			"elasticstack_fleet_output":             fleet.ResourceOutput(),
-			"elasticstack_fleet_server_host":        fleet.ResourceFleetServerHost(),
-			"elasticstack_fleet_integration":        fleet.ResourceIntegration(),
-			"elasticstack_fleet_integration_policy": fleet.ResourceIntegrationPolicy(),
 		},
 	}